feat: extracted callback data logic into separate package
